feat: archive -> news 이름 변경
