--- src/main/java/com/sejong/archiveservice/application/archive/dto/ArchiveReqDto.java
@@ -1,14 +1,0 @@
-package com.sejong.archiveservice.application.archive.dto;
-
-import java.util.List;
-
-public record ArchiveReqDto(
-    String title,
-    String summary,
-    String content,
-    String category,
-    String writerId,
-    List<String> participantIds,
-    List<String> tags
-) {
-}
