--- src/main/java/com/sejong/archiveservice/application/archive/dto/UpdateFileInfoReqDto.java
@@ -1,6 +1,0 @@
-package com.sejong.archiveservice.application.archive.dto;
-
-public record UpdateFileInfoReqDto(
-        String thumbnailPath
-) {
-}
