+++ src/main/java/com/sejong/archiveservice/application/internal/NewsInternalController.java
@@ -1,0 +1,23 @@
+package com.sejong.archiveservice.application.internal;
+
+import com.sejong.archiveservice.application.news.service.NewsService;
+import lombok.RequiredArgsConstructor;
+import org.springframework.http.ResponseEntity;
+import org.springframework.web.bind.annotation.GetMapping;
+import org.springframework.web.bind.annotation.PathVariable;
+import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+@RequiredArgsConstructor
+@RequestMapping("/internal/archive")
+public class NewsInternalController {
+
+    private final NewsService newsService;
+
+    @GetMapping("/check/{archiveId}")
+    public ResponseEntity<Boolean> checkArchiveId(@PathVariable("archiveId") Long archiveId) {
+        Boolean response = newsService.exists(archiveId);
+        return ResponseEntity.ok(response);
+    }
+}
