+++ src/main/java/com/sejong/archiveservice/application/news/dto/NewsReqDto.java
@@ -1,0 +1,14 @@
+package com.sejong.archiveservice.application.news.dto;
+
+import java.util.List;
+
+public record NewsReqDto(
+    String title,
+    String summary,
+    String content,
+    String category,
+    String writerId,
+    List<String> participantIds,
+    List<String> tags
+) {
+}
