+++ src/main/java/com/sejong/archiveservice/application/news/dto/UpdateFileInfoReqDto.java
@@ -1,0 +1,6 @@
+package com.sejong.archiveservice.application.news.dto;
+
+public record UpdateFileInfoReqDto(
+        String thumbnailPath
+) {
+}
